refactor(graph-tools): split collapseBoundary into FCIS pipeline modules
Apply fp-rearchitecting Pattern 5 (pipeline composition) + M1
(deep-narrow module) to view/collapseBoundary.ts (was 622 lines,
cyclomatic complexity 85).

The public API was already a 4-step pipeline (normalize → folder/louvain
candidates → greedy selection); each step is now its own internal module
with one concern:

- collapseBoundary/index.ts     deep-narrow public entry
- collapseBoundary/types.ts     shared types (no logic)
- collapseBoundary/normalize.ts NormalizedGraph + protectedIds + pageRank
- collapseBoundary/folder.ts    folder-first candidate generation
- collapseBoundary/louvain.ts   Louvain candidate generation
- collapseBoundary/selection.ts greedy selection + cluster stats

Per-module complexity ~15-25 each; public API surface unchanged
(same exports, same behaviour). All 15 collapse tests pass.

Why now: the vault->projectRoot rename added +1 churn event to ~16
graph-tools files, raising package-avg-turbulence from 48.85 to 52.68
and pushing pressure-axes RSCD over 1.0. Splitting the heaviest file
restores headroom by distributing complexity across modules.

Per brain/workflows/engineering/architectural-complexity/fp-rearchitecting.
